refactor(api): extract DataSourceService from datasource/[inventoryId] route
