+++ app/src/app/api/v0/datasource/[inventoryId]/route.ts
@@ -1,18 +1,14 @@
+import DataSourceService from "@/backend/DataSourceService";
 import { filterSources } from "@/lib/filter-sources";
 import { db } from "@/models";
 import { City } from "@/models/City";
 import { DataSource } from "@/models/DataSource";
-import { Inventory } from "@/models/Inventory";
 import { Scope } from "@/models/Scope";
 import { SubCategory } from "@/models/SubCategory";
 import { SubCategoryValue } from "@/models/SubCategoryValue";
 import { SubSector } from "@/models/SubSector";
-import {
-  SubSectorValue,
-  SubSectorValueCreationAttributes,
-} from "@/models/SubSectorValue";
+import { SubSectorValue } from "@/models/SubSectorValue";
 import { apiHandler } from "@/util/api";
-import { randomUUID } from "crypto";
 import createHttpError from "http-errors";
 import { NextRequest, NextResponse } from "next/server";
 import { Op } from "sequelize";
@@ -92,9 +88,18 @@ export const POST = apiHandler(async (req: NextRequest, { params }) => {
     where: { datasourceId: body.dataSourceIds },
     include: [
       { model: SubSector, required: false, as: "subSector" },
-      { model: SubCategory, required: false, as: "subCategory", include: [{
-        model: SubSector, required: false, as: "subsector",
-      }]},
+      {
+        model: SubCategory,
+        required: false,
+        as: "subCategory",
+        include: [
+          {
+            model: SubSector,
+            required: false,
+            as: "subsector",
+          },
+        ],
+      },
     ],
   });
   if (!sources) {
@@ -118,7 +123,10 @@ export const POST = apiHandler(async (req: NextRequest, { params }) => {
       const result = { id: source.datasourceId, success: true };
 
       if (source.retrievalMethod === "global_api") {
-        result.success = await retrieveGlobalAPISource(source, inventory);
+        result.success = await DataSourceService.applyGlobalAPISource(
+          source,
+          inventory,
+        );
       } else {
         console.error(
           `Unsupported retrieval method ${source.retrievalMethod} for data source ${source.datasourceId}`,
@@ -142,142 +150,3 @@ export const POST = apiHandler(async (req: NextRequest, { params }) => {
   });
 });
 
-async function initSubSectorValue(
-  source: DataSource,
-  inventory: Inventory,
-  totalEmissions: number,
-  values: Partial<SubSectorValueCreationAttributes>,
-  sectorId: string,
-  subsectorId: string,
-): Promise<SubSectorValue> {
-  if (!sectorId) {
-    throw new createHttpError.InternalServerError(
-      "Failed to find sector ID for source " + source.datasourceId,
-    );
-  }
-  if (!subsectorId) {
-    throw new createHttpError.InternalServerError(
-      "Failed to find subsector ID for source " + source.datasourceId,
-    );
-  }
-
-  let sectorValue = await db.models.SectorValue.findOne({
-    where: {
-      sectorId,
-      inventoryId: inventory.inventoryId,
-    },
-  });
-  // TODO have to init/ update totalEmissions here?
-  if (!sectorValue) {
-    sectorValue = await db.models.SectorValue.create({
-      sectorValueId: randomUUID(),
-      sectorId,
-      inventoryId: inventory.inventoryId,
-      totalEmissions,
-    });
-  } else {
-    await sectorValue.update({
-      totalEmissions: (sectorValue.totalEmissions || 0) + totalEmissions,
-    });
-  }
-  const subSectorValue = await db.models.SubSectorValue.create({
-    ...values,
-    sectorValueId: sectorValue.sectorValueId,
-    subsectorId,
-    subsectorValueId: randomUUID(),
-  });
-  return subSectorValue;
-}
-
-async function retrieveGlobalAPISource(
-  source: DataSource,
-  inventory: Inventory,
-): Promise<boolean> {
-  const referenceNumber =
-    source.subCategory?.referenceNumber || source.subSector?.referenceNumber;
-
-  if (
-    !source.apiEndpoint ||
-    !inventory.city.locode ||
-    inventory.year == null ||
-    !(source.subsectorId || source.subcategoryId) ||
-    !referenceNumber
-  ) {
-    return false;
-  }
-
-  const url = source.apiEndpoint
-    .replace("openearth.cloud", "openearth.dev") // TODO remove once data catalogue is fixed
-    .replace(":locode", inventory.city.locode.replace("-", " "))
-    .replace(":year", inventory.year.toString())
-    .replace(":gpcReferenceNumber", referenceNumber);
-
-  let data;
-  try {
-    const response = await fetch(url);
-    data = await response.json();
-  } catch (err) {
-    console.error(
-      `Failed to query data source ${source.datasourceId} at URL ${url}:`,
-      err,
-    );
-    return false;
-  }
-
-  if (typeof data.totals !== "object") {
-    console.error("Incorrect response from Global API for URL:", url, data);
-    return false;
-  }
-
-  const emissions = data.totals.emissions;
-  // TODO store values for co2, ch4, n2o separately for accounting and editing
-  const totalEmissions = emissions.co2eq_100yr;
-  const values = {
-    datasourceId: source.datasourceId,
-    totalEmissions,
-    inventoryId: inventory.inventoryId,
-  };
-
-  if (source.subsectorId) {
-    await initSubSectorValue(
-      source,
-      inventory,
-      totalEmissions,
-      values,
-      source.subSector.sectorId!,
-      source.subsectorId,
-    );
-  } else if (source.subcategoryId) {
-    // add parent SubSectorValue if not present yet
-    let subSectorValue = await db.models.SubSectorValue.findOne({
-      where: {
-        subsectorId: source.subCategory?.subsectorId,
-        inventoryId: inventory.inventoryId,
-      },
-    });
-    if (!subSectorValue) {
-      subSectorValue = await initSubSectorValue(
-        source,
-        inventory,
-        totalEmissions,
-        values,
-        source.subCategory?.subsector?.sectorId!,
-        source.subCategory?.subsectorId!,
-      );
-    } else {
-      await subSectorValue.update({
-        totalEmissions: (subSectorValue.totalEmissions || 0) + totalEmissions,
-      });
-    }
-    const subCategoryValue = await db.models.SubCategoryValue.create({
-      ...values,
-      subcategoryValueId: randomUUID(),
-      subcategoryId: source.subcategoryId,
-      subsectorValueId: subSectorValue.subsectorValueId,
-    });
-  } else {
-    return false;
-  }
-
-  return true;
-}
