+++ app/src/backend/DataSourceService.ts
@@ -1,0 +1,159 @@
+import { db } from "@/models";
+import { DataSource } from "@/models/DataSource";
+import { Inventory } from "@/models/Inventory";
+import {
+  SubSectorValue,
+  SubSectorValueCreationAttributes,
+} from "@/models/SubSectorValue";
+import { randomUUID } from "crypto";
+import createHttpError from "http-errors";
+
+export default class DataSourceService {
+  public static async retrieveGlobalAPISource(
+    source: DataSource,
+    inventory: Inventory,
+  ): Promise<any | null> {
+    const referenceNumber =
+      source.subCategory?.referenceNumber || source.subSector?.referenceNumber;
+    if (
+      !source.apiEndpoint ||
+      !inventory.city.locode ||
+      inventory.year == null ||
+      !(source.subsectorId || source.subcategoryId) ||
+      !referenceNumber
+    ) {
+      return false;
+    }
+
+    const url = source.apiEndpoint
+      .replace(":locode", inventory.city.locode.replace("-", " "))
+      .replace(":year", inventory.year.toString())
+      .replace(":gpcReferenceNumber", referenceNumber);
+
+    let data;
+    try {
+      const response = await fetch(url);
+      data = await response.json();
+    } catch (err) {
+      console.error(
+        `Failed to query data source ${source.datasourceId} at URL ${url}:`,
+        err,
+      );
+      return false;
+    }
+
+    if (typeof data.totals !== "object") {
+      console.error("Incorrect response from Global API for URL:", url, data);
+      return false;
+    }
+  }
+
+  public static async applyGlobalAPISource(
+    source: DataSource,
+    inventory: Inventory,
+  ): Promise<boolean> {
+    const data = await DataSourceService.retrieveGlobalAPISource(
+      source,
+      inventory,
+    );
+
+    const emissions = data.totals.emissions;
+    // TODO store values for co2, ch4, n2o separately for accounting and editing
+    const totalEmissions = emissions.co2eq_100yr;
+    const values = {
+      datasourceId: source.datasourceId,
+      totalEmissions,
+      inventoryId: inventory.inventoryId,
+    };
+
+    if (source.subsectorId) {
+      await DataSourceService.initSubSectorValue(
+        source,
+        inventory,
+        totalEmissions,
+        values,
+        source.subSector.sectorId!,
+        source.subsectorId,
+      );
+    } else if (source.subcategoryId) {
+      // add parent SubSectorValue if not present yet
+      let subSectorValue = await db.models.SubSectorValue.findOne({
+        where: {
+          subsectorId: source.subCategory?.subsectorId,
+          inventoryId: inventory.inventoryId,
+        },
+      });
+      if (!subSectorValue) {
+        subSectorValue = await DataSourceService.initSubSectorValue(
+          source,
+          inventory,
+          totalEmissions,
+          values,
+          source.subCategory?.subsector?.sectorId!,
+          source.subCategory?.subsectorId!,
+        );
+      } else {
+        await subSectorValue.update({
+          totalEmissions: (subSectorValue.totalEmissions || 0) + totalEmissions,
+        });
+      }
+      const subCategoryValue = await db.models.SubCategoryValue.create({
+        ...values,
+        subcategoryValueId: randomUUID(),
+        subcategoryId: source.subcategoryId,
+        subsectorValueId: subSectorValue.subsectorValueId,
+      });
+    } else {
+      return false;
+    }
+
+    return true;
+  }
+
+  private static async initSubSectorValue(
+    source: DataSource,
+    inventory: Inventory,
+    totalEmissions: number,
+    values: Partial<SubSectorValueCreationAttributes>,
+    sectorId: string,
+    subsectorId: string,
+  ): Promise<SubSectorValue> {
+    if (!sectorId) {
+      throw new createHttpError.InternalServerError(
+        "Failed to find sector ID for source " + source.datasourceId,
+      );
+    }
+    if (!subsectorId) {
+      throw new createHttpError.InternalServerError(
+        "Failed to find subsector ID for source " + source.datasourceId,
+      );
+    }
+
+    let sectorValue = await db.models.SectorValue.findOne({
+      where: {
+        sectorId,
+        inventoryId: inventory.inventoryId,
+      },
+    });
+    // TODO have to init/ update totalEmissions here?
+    if (!sectorValue) {
+      sectorValue = await db.models.SectorValue.create({
+        sectorValueId: randomUUID(),
+        sectorId,
+        inventoryId: inventory.inventoryId,
+        totalEmissions,
+      });
+    } else {
+      await sectorValue.update({
+        totalEmissions: (sectorValue.totalEmissions || 0) + totalEmissions,
+      });
+    }
+    const subSectorValue = await db.models.SubSectorValue.create({
+      ...values,
+      sectorValueId: sectorValue.sectorValueId,
+      subsectorId,
+      subsectorValueId: randomUUID(),
+    });
+    return subSectorValue;
+  }
+}
